+++ AoC2019/Solutions/AoC_5.h
@@ -3,10 +3,10 @@
 
 #include <vector>
 
-#include "../Classes/IntCodeCPU/IntCodeCPU.h"
+#include "Classes/IntCodeCPU/IntCodeCPU.h"
 
 namespace AoC_5 {
-    long long AB(std::vector<long long> program, int systemID) {
+    long long AB(std::vector<long long> program, long long systemID) {
         IntCodeCPU computer(program);
         computer.Input(systemID);
         computer.RunToEnd();
